+++ pages/api/peer_suggestion_item/[id]/index.ts
@@ -70,6 +70,25 @@ export default async function handler(
 
       const { id } = validator.cast(req.query);
 
+      const submittedPeerSuggestion =
+        await prisma.submittedPeerSuggestion.findFirst({
+          where: {
+            PeerSuggestion: {
+              PeerSuggestionItem: {
+                some: {
+                  id: {
+                    equals: id as string,
+                  },
+                },
+              },
+            },
+          },
+        });
+
+      if (submittedPeerSuggestion) {
+        throw new Error("Peer Suggestion is already submitted");
+      }
+
       const peerSuggestionItem = await prisma.peerSuggestionItem.delete({
         where: {
           id,
@@ -104,6 +123,25 @@ export default async function handler(
       const { actionTaken, remarks, suggestion, pageNumber } = validator.cast(
         req.body
       );
+
+      const submittedPeerSuggestion =
+        await prisma.submittedPeerSuggestion.findFirst({
+          where: {
+            PeerSuggestion: {
+              PeerSuggestionItem: {
+                some: {
+                  id: {
+                    equals: id as string,
+                  },
+                },
+              },
+            },
+          },
+        });
+
+      if (submittedPeerSuggestion) {
+        throw new Error("Peer Suggestion is already submitted");
+      }
 
       const peerSuggestionItem = await prisma.peerSuggestionItem.update({
         where: {
